fix: исправлены критические ошибки в backend
1. Исправлен тип данных telegram_id с INTEGER на BIGINT
   - Создана миграция для изменения типа в базе данных
   - Теперь поддерживаются большие Telegram ID (например, 5673644913)
   - Решена ошибка 'integer out of range'

2. Исправлена работа с датами в tune уведомлениях
   - Заменен datetime.utcnow() на datetime.now(timezone.utc)
   - Решена ошибка 'can't subtract offset-naive and offset-aware datetimes'
   - Добавлен импорт timezone

Оба исправления протестированы и работают корректно.

+++ backend/app/models/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, JSON, Boolean
+from sqlalchemy import Column, Integer, String, DateTime, JSON, Boolean, BigInteger
 from sqlalchemy.sql import func
 from sqlalchemy.orm import relationship
 
@@ -9,7 +9,7 @@ class User(Base):
     __tablename__ = "users"
 
     id = Column(Integer, primary_key=True, index=True)
-    telegram_id = Column(Integer, unique=True, index=True, nullable=False)
+    telegram_id = Column(BigInteger, unique=True, index=True, nullable=False)
     first_name = Column(String, nullable=False)
     last_name = Column(String, nullable=True)
     username = Column(String, nullable=True)
